App and course completed

+++ app/src/main/java/com/ys_production/tasktimer/MainActivityFragment.kt
@@ -6,9 +6,13 @@ import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModelProvider
+import androidx.fragment.app.activityViewModels
+import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import com.ys_production.tasktimer.databinding.FragmentMainActivityBinding
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -17,8 +21,8 @@ import kotlinx.coroutines.launch
 private const val TAG = "MainActivityFragment"
 
 class MainActivityFragment : Fragment(), CursorAdapter.OnTaskClickListener {
-    private val bindding by lazy { FragmentMainActivityBinding.inflate(layoutInflater) }
-    private val viewModel by lazy { ViewModelProvider(this)[TaskTimerViewModel::class.java] }
+    private val binding by lazy { FragmentMainActivityBinding.inflate(layoutInflater) }
+    private val viewModel: TaskTimerViewModel by activityViewModels()
     private val adapter = CursorAdapter(null, this)
 
     interface OnTaskEdit {
@@ -28,20 +32,43 @@ class MainActivityFragment : Fragment(), CursorAdapter.OnTaskClickListener {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
     ): View {
-        bindding.ffRecyclerView.layoutManager = LinearLayoutManager(requireContext())
-        bindding.ffRecyclerView.adapter = adapter
+        binding.ffRecyclerView.layoutManager = LinearLayoutManager(requireContext())
+        binding.ffRecyclerView.adapter = adapter
+        val itemSwipeHelper =
+            ItemTouchHelper(object : ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT) {
+                override fun onMove(
+                    recyclerView: RecyclerView,
+                    viewHolder: RecyclerView.ViewHolder,
+                    target: RecyclerView.ViewHolder
+                ): Boolean {
+                    return false
+                }
+
+                override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
+                    if (direction == ItemTouchHelper.LEFT) {
+                        val task = (viewHolder as CursorViewHolder).task
+                        if (viewModel.editedTaskId == task.id.toLong()){
+                            adapter.notifyItemChanged(viewHolder.adapterPosition)
+                            Toast.makeText(context,getString(R.string.can_not_delete_toast), Toast.LENGTH_LONG).show()
+                        }else {
+                            onDeleteClick(task, viewHolder.adapterPosition)
+                        }
+                    }
+                }
+            })
+        itemSwipeHelper.attachToRecyclerView(binding.ffRecyclerView)
         viewModel.cursor.observe(viewLifecycleOwner) {
             val cursor = adapter.swapCursor(it)
-            CoroutineScope(Dispatchers.Default).launch { cursor?.close() }
+            CoroutineScope(Dispatchers.IO).launch { cursor?.close() }
         }
-        viewModel.timing.observe(viewLifecycleOwner){
-            bindding.itemSelected.text = if (it != null){
-                getString(R.string.currently_timing,it)
-            }else{
+        viewModel.timing.observe(viewLifecycleOwner) {
+            binding.itemSelected.text = if (it != null) {
+                getString(R.string.currently_timing, it)
+            } else {
                 getString(R.string.no_task_selected)
             }
         }
-        return bindding.root
+        return binding.root
     }
 
     override fun onAttach(context: Context) {
@@ -53,9 +80,15 @@ class MainActivityFragment : Fragment(), CursorAdapter.OnTaskClickListener {
         (activity as OnTaskEdit?)?.onTaskEdit(task)
     }
 
-    override fun onDeleteClick(task: Task) {
+    fun onDeleteClick(task: Task, position: Int) {
         Log.d(TAG, "onDeleteClick: ")
-        viewModel.deleteTask(task)
+        AlertDialog.Builder(requireContext())
+            .setMessage(getString(R.string.delete_task_massage, task.name))
+            .setPositiveButton("Delete") { _, _ ->
+                viewModel.deleteTask(task)
+            }.setNegativeButton("Cancel") { _, _ ->
+                adapter.notifyItemChanged(position)
+            }.create().show()
     }
 
     override fun onTaskBackLongClick(task: Task) {
